+++ alliance/src/asimut/src/vh_debug.c
@@ -36,7 +36,7 @@ char         **str   ;			/* recognized strings		*/
   int                    i;
   struct chain	        *ptr_abl;
   static char           *buffer      = NULL;
-  static          int    buffer_size = 0;
+  static          long   buffer_size = 0;
   static struct circuit *circuit_pnt = NULL;
 
   struct chain   *chain_pnt;
@@ -1902,7 +1902,7 @@ char *type;				/* structure's type		*/
             {
             size = siz [(typ [idx] & TYPE_DFN)];
             pntr = (void *)
-                   (((unsigned int) pnt [idx].dat) + (size * nmbrs [1]));
+                   (((unsigned long) pnt [idx].dat) + (size * nmbrs [1]));
             push (stk, &stkpnt, pntr, pshtype);
             readflg = 1;
             }
@@ -1919,7 +1919,7 @@ char *type;				/* structure's type		*/
             {
             size = sizeof (void *);
             pntr = * (void **)
-                     (((unsigned int) pnt [idx].dat) + (size * nmbrs [1]));
+                     (((unsigned long) pnt [idx].dat) + (size * nmbrs [1]));
             push (stk, &stkpnt, pntr, pshtype);
             readflg = 1;
             }
